backend/requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
twilio
pydantic